pyproject.toml
[tool.poetry]
name = "botsecretary"
version = "0.1.0"
description = "Personal AI Bot Assistent"
authors = ["Aleksandr Kuznetsov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.2"
pydantic = {extras = ["dotenv"], version = "^1.10.9"}
openai = "^0.27.8"
python-telegram-bot = "^20.3"
jinja2 = "^3.1.2"
loguru = "^0.7.0"
sqlalchemy = "^2.0.16"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
